+++ packages/remix-cloudflare/index.ts
@@ -12,6 +12,8 @@ export {
 export {
   createRequestHandler,
   createSession,
+  unstable_defineLoader,
+  unstable_defineAction,
   defer,
   broadcastDevReady,
   logDevReady,